# server/swagger.yaml
openapi: 3.0.3
info:
  title: Yasuo API - Yasup Shop
  description: |-
    This API provides programmatic access to a vast collection of merchandise related to the popular game League of Legends.
    Integrate this API into your application to offer users a convenient way to browse, search, and potentially purchase officially licensed League of Legends merchandise.

  contact:
    email: nhockkutean2@gmail.com
  version: 1.0.0
...
servers:
  - url: http://127.0.0.1:1337
  - url: https://yasuo-api-serverless.netlify.app
...
paths:
...
  /api/v1/users/activate:
    patch:
      tags:
        - Users
      summary: Activate User Account
      description: Activates a user account using the provided activation code.
      operationId: activate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  $ref: '#/components/schemas/Email'
                code:
                  $ref: '#/components/schemas/Code'
      responses:
        '200':
          description: User account successfully activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  message:
                    type: string
                    description: 'A message indicating the outcome of the account activation process'
                    example: 'Activate account successfully.'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/InvalidOrExpiredToken'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/users/forgot-password/{email}:
    get:
      tags:
        - Users
      summary: Forgot Password
      description: Sends a forgot password code to the user's email.
      operationId: forgotPassword
      parameters:
        - $ref: '#/components/parameters/EmailParam'
      responses:
        '200':
          description: Forgot password code successfully sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  message:
                    type: string
                    description: 'A message indicating the outcome of the forgot password code request'
                    example: 'Forgot password code has been sent to your email. Please check.'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/users/reset-password:
    patch:
      tags:
        - Users
      summary: Reset User Password
      description: Resets a user's password using the provided email, reset code, and new password.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  $ref: '#/components/schemas/Email'
                code:
                  $ref: '#/components/schemas/Code'
                newPassword:
                  $ref: '#/components/schemas/Password'
      responses:
        '200':
          description: Password successfully reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  message:
                    type: string
                    description: 'A message indicating the outcome of the password reset process'
                    example: 'Reset password successfully.'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/InvalidOrExpiredToken'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/users/update-password:
    patch:
      tags:
        - Users
      summary: Update user password
      description: Updates the password of the authenticated user.
      operationId: updatePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  $ref: '#/components/schemas/Password'
                newPassword:
                  type: string
                  description: 'The new password to be set'
                  example: 'userpassword3Mm$2'
                passwordConfirm:
                  type: string
                  description: 'Confirmation of the new password'
                  example: 'userpassword3Mm$2'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  message:
                    type: string
                    description: 'A message indicating the success of the operation'
                    example: 'Update password successfully!'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/users/ban-account:
    post:
      tags:
        - Users
      summary: Ban a user account - Admin
      description: |
        Bans a user account based on the provided email.

        **Only Admin can perform this action**
      operationId: banAccount
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: 'The email address of the user to ban'
                  example: 'tester@gmail.com'
      responses:
        '200':
          description: User account banned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  message:
                    type: string
                    description: 'A message indicating the success of the operation'
                    example: 'User has been banned!'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/users/me:
    get:
      tags:
        - Users
      summary: Get user's own data
      description: Retrieves information about the currently authenticated user.
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    patch:
      tags:
        - Users
      summary: Update user profile
      description: Updates the profile information of the authenticated user.
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        description: This route can only update the name, the photo, or both of the user.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "The user's name"
                  example: 'Tester'
                photo:
                  type: string
                  description: "The path to the user's profile photo"
                  example: '/img/users/default.png'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  message:
                    type: string
                    description: 'A message indicating the success of the operation'
                    example: 'Update user successfully.'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/users:
    get:
      tags:
        - Users
      summary: Get all users - Admin
      description: |
        Retrieves a list of all users.

        **Only Admin can perform this action**
      operationId: getAllUsers
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: additions
          description: |
            Advanced filtering, sorting, projecting, pagination.

            Filtering: { "name": "Tester" }

            Sorting: { "sort": "name" }

            Projecting: { "fields": "name" }

            Pagination: { "page": 1, "limit": 3 }
          required: false
          schema:
            type: object
            additionalProperties: true
      responses:
        '200':
          description: Users retrieved successfully - User data can have more fileds than the example value.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  numResults:
                    $ref: '#/components/schemas/NumResults'
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags:
        - Users
      summary: Create a new user - Admin
      description: |
        Creates a new user account with the provided details.

        **Only Admin can perform this action**
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        description: The request body can have more fields based on the admin's needs.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'The name of the user'
                  example: 'Tester'
                email:
                  $ref: '#/components/schemas/Email'
                password:
                  $ref: '#/components/schemas/Password'
                active:
                  type: boolean
                  description: 'The active status of the user'
                  example: true
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                  message:
                    type: string
                    description: 'A message indicating the success of the operation'
                    example: 'Create user successfully.'
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/v1/users/{email}:
    get:
      tags:
        - Users
      summary: Get user by email - Admin
      description: |
        Retrieves the details of a user by their email address.

        **Only Admin can perform this action**
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmailParam'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/SuccessMessage'
                    numResults:
                      $ref: '#/components/schemas/NumResults'
                    user:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags:
        - Users
      summary: Delete user by email
      description: |
        Deletes a user by their email address.

        **An admin can delete any user.
        Users can only delete their own accounts.
        Deleting does not actually remove the user; it just marks the user as deleted.**
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmailParam'
      responses:
        '204':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    SuccessMessage:
      type: string
      description: The status of the response.
      example: success
    FailMessage:
      type: string
      description: The status of the response.
      example: fail
    ErrorMessage:
      type: string
      description: The status of the response.
      example: error
...
    NumResults:
      type: integer
      description: The number of entities returned.
      example: 1
...
    User:
      type: object
      properties:
        _id:
          type: string
          format: MongoID
          description: 'The unique identifier for the user'
          example: '6644461874a0a872f25dbc1c'
        name:
          type: string
          description: 'The name of the user'
          example: 'Tester'
        email:
          $ref: '#/components/schemas/Email'
        photo:
          type: string
          description: "The URL of the user's profile photo"
          example: '/img/users/default.png'
        active:
          type: boolean
          description: "Indicates whether the user's account is active (Use code received in user's email to activate user)"
          example: false
        role:
          type: string
          description: 'The role assigned to the user'
          example: 'user'
        createdAt:
          type: string
          format: date-time
          description: 'The timestamp when the user account was created'
          example: '2024-05-15T05:12:46.943Z'
        updatedAt:
          type: string
          format: date-time
          description: 'The timestamp when the user account was last updated'
          example: '2024-05-15T05:12:47.286Z'
    Email:
      type: string
      description: 'The email address of the user'
      example: 'tester@gmail.com'
    Code:
      type: string
      description: "The code sent to the user's email"
      example: '0b0ada25c3ba'
    Password:
      type: string
      description: 'The new password for the user account (Password must contain at least one lowercase letter, one uppercase letter, one number, and one special character).'
      example: 'userpassword3Mm$'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    EmailParam:
      in: path
      name: email
      required: true
      schema:
        type: string
      description: The email address of the user.
      example: 'tester@gmail.com'
  responses:
    BadRequest:
      description: Bad request!
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/FailMessage'
              message:
                type: string
                description: The error message explaining why the request failed.
                example: Something went wrong! (This can be any string based on the received error)
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/FailMessage'
              message:
                type: string
                description: The error message explaining that the entity was not found.
                example: Entity not found!
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/ErrorMessage'
              message:
                type: string
                example: Something went wrong!
    InvalidOrExpiredToken:
      description: Invalid or Expired Token
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/FailMessage'
              message:
                type: string
                description: The error message explaining why the request failed.
                example: Invalid token or token has expired!
    Unauthenticated:
      description: Unauthenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/FailMessage'
              message:
                type: string
                description: 'A message indicating that the user is unauthenticated'
                example: 'Unauthenticated! (This can be any string based on the received error)'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/FailMessage'
              message:
                type: string
                description: 'A message indicating that the user is unauthorized'
                example: 'Unauthorized! (This can be any string based on the received error)'